WIP: Refactor CommonSpace
