+++ src/util/heap/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod chunk_map;
 pub(crate) mod externalpageresource;
 pub(crate) mod freelistpageresource;
 pub(crate) mod gc_trigger;
-mod heap_meta;
+pub(crate) mod heap_meta;
 pub(crate) mod monotonepageresource;
 pub(crate) mod pageresource;
 pub(crate) mod space_descriptor;
@@ -15,7 +15,6 @@ mod vmrequest;
 pub(crate) use self::accounting::PageAccounting;
 pub(crate) use self::blockpageresource::BlockPageResource;
 pub(crate) use self::freelistpageresource::FreeListPageResource;
-pub(crate) use self::heap_meta::HeapMeta;
 pub use self::layout::vm_layout;
 pub(crate) use self::monotonepageresource::MonotonePageResource;
 pub(crate) use self::pageresource::PageResource;
